/adminから/admin/dashboardに変更
