+++ app/(public)/admin/events/page.tsx
@@ -24,7 +24,7 @@ export default function AdminEventsPage() {
     <div className="min-h-screen bg-gray-50 pb-20">
       <div className="bg-white shadow-sm">
         <div className="mx-auto max-w-7xl px-4 py-6">
-          <Link href="/admin" className="text-sm text-blue-600 hover:text-blue-800">
+          <Link href="/admin/dashboard" className="text-sm text-blue-600 hover:text-blue-800">
             ← ダッシュボードに戻る
           </Link>
           <h1 className="mt-2 text-3xl font-bold text-gray-900">イベント管理</h1>
